+++ public/index.html
@@ -29,8 +29,9 @@
             href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css"
     />
     <link
+            href="https://fonts.googleapis.com/icon?family=Material+Icons"
             rel="stylesheet"
-            href="https://fonts.googleapis.com/icon?family=Material+Icons"/>
+    />
     <title>React App</title>
   </head>
   <body>
