+++ src/components/Navbar.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
+import { NavLink } from "react-router-dom";
 
 export  const NavBar: React.FC = () => (
     <nav>
         <div className="nav-wrapper purple darken-2 px1">
             <a href="/" className="brand-logo">React + Typescript</a>
                 <ul className="right hide-on-med-and-down">
-                    <li><a href="/">Список дел</a></li>
-                    <li><a href="/">Информация</a></li>
+                    <li><NavLink to="/">Список дел</NavLink></li>
+                    <li><NavLink to="/about">Информация</NavLink></li>
                 </ul>
         </div>
     </nav>
