+++ src/index.css
@@ -1,23 +1,26 @@
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
+  'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+  sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
 
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
+  monospace;
 }
+
 .px1 {
   padding-left: 1rem;
   padding-right: 1rem;
 }
+
 .mt2 {
   margin-top: 2rem;
 }
+
 .todo {
   margin-bottom: 1rem;
 }
@@ -25,6 +28,7 @@ code {
 .todo.completed span {
   text-decoration: line-through;
 }
+
 .todo label {
   width: 100%;
   display: flex;
@@ -33,8 +37,8 @@ code {
   cursor: pointer;
   padding: 1rem;
   background: #f3f3f3;
-
 }
+
 .todo i {
   z-index: 2;
 }
